save dev state — zcash extras
