+++ Cargo.toml
@@ -5,6 +5,7 @@ members = [
     "zcash_client_backend",
     "zcash_client_sqlite",
     "zcash_extensions",
+    "zcash_extras",
     "zcash_history",
     "zcash_primitives",
     "zcash_proofs",
