+++ zcash_client_sqlite/Cargo.toml
@@ -27,7 +27,9 @@ libsqlite3-sys= { version = "0.25.2", features = ["bundled"] }
 time = "0.3"
 tokio = { version = "1.20", features = ["full"] }
 zcash_client_backend = { version = "0.5", path = "../zcash_client_backend" }
+zcash_extras = { version = "0.1", path = "../zcash_extras" }
 zcash_primitives = { version = "0.5", path = "../zcash_primitives" }
+
 
 [dev-dependencies]
 rand_core = "0.5.1"
