+++ zcash_client_sqlite/src/with_async/mod.rs
@@ -1,12 +1,10 @@
 pub mod wallet_actions;
 
-use std::cmp;
 use std::collections::HashMap;
-use std::fmt::Debug;
 use std::path::Path;
-use zcash_client_backend::data_api::wallet::ANCHOR_OFFSET;
 use zcash_client_backend::data_api::{PrunedBlock, ReceivedTransaction, SentTransaction};
 use zcash_client_backend::wallet::{AccountId, SpendableNote};
+use zcash_extras::{WalletRead, WalletWrite};
 use zcash_primitives::block::BlockHash;
 use zcash_primitives::consensus::BlockHeight;
 use zcash_primitives::memo::Memo;
@@ -24,182 +22,6 @@ where
     tokio::task::spawn_blocking(blocking_fn)
         .await
         .expect("spawn_blocking to succeed")
-}
-
-#[async_trait::async_trait]
-pub trait WalletRead: Send + Sync + 'static {
-    type Error;
-    type NoteRef: Copy + Debug;
-    type TxRef: Copy + Debug;
-
-    /// Returns the minimum and maximum block heights for stored blocks.
-    ///
-    /// This will return `Ok(None)` if no block data is present in the database.
-    async fn block_height_extrema(&self)
-        -> Result<Option<(BlockHeight, BlockHeight)>, Self::Error>;
-
-    /// Returns the default target height (for the block in which a new
-    /// transaction would be mined) and anchor height (to use for a new
-    /// transaction), given the range of block heights that the backend
-    /// knows about.
-    ///
-    /// This will return `Ok(None)` if no block data is present in the database.
-    async fn get_target_and_anchor_heights(
-        &self,
-    ) -> Result<Option<(BlockHeight, BlockHeight)>, Self::Error> {
-        self.block_height_extrema().await.map(|heights| {
-            heights.map(|(min_height, max_height)| {
-                let target_height = max_height + 1;
-
-                // Select an anchor ANCHOR_OFFSET back from the target block,
-                // unless that would be before the earliest block we have.
-                let anchor_height = BlockHeight::from(cmp::max(
-                    u32::from(target_height).saturating_sub(ANCHOR_OFFSET),
-                    u32::from(min_height),
-                ));
-
-                (target_height, anchor_height)
-            })
-        })
-    }
-
-    /// Returns the block hash for the block at the given height, if the
-    /// associated block data is available. Returns `Ok(None)` if the hash
-    /// is not found in the database.
-    async fn get_block_hash(
-        &self,
-        block_height: BlockHeight,
-    ) -> Result<Option<BlockHash>, Self::Error>;
-
-    /// Returns the block hash for the block at the maximum height known
-    /// in stored data.
-    ///
-    /// This will return `Ok(None)` if no block data is present in the database.
-    async fn get_max_height_hash(&self) -> Result<Option<(BlockHeight, BlockHash)>, Self::Error> {
-        println!("before extrema");
-        let extrema = self.block_height_extrema().await?;
-        println!("extrema {extrema:?}");
-        let res = if let Some((_, max_height)) = extrema {
-            self.get_block_hash(max_height)
-                .await
-                .map(|hash_opt| hash_opt.map(move |hash| (max_height, hash)))?
-        } else {
-            None
-        };
-
-        Ok(res)
-    }
-
-    /// Returns the block height in which the specified transaction was mined,
-    /// or `Ok(None)` if the transaction is not mined in the main chain.
-    async fn get_tx_height(&self, txid: TxId) -> Result<Option<BlockHeight>, Self::Error>;
-
-    /// Returns the payment address for the specified account, if the account
-    /// identifier specified refers to a valid account for this wallet.
-    ///
-    /// This will return `Ok(None)` if the account identifier does not correspond
-    /// to a known account.
-    async fn get_address(&self, account: AccountId) -> Result<Option<PaymentAddress>, Self::Error>;
-
-    /// Returns all extended full viewing keys known about by this wallet.
-    async fn get_extended_full_viewing_keys(
-        &self,
-    ) -> Result<HashMap<AccountId, ExtendedFullViewingKey>, Self::Error>;
-
-    /// Checks whether the specified extended full viewing key is
-    /// associated with the account.
-    async fn is_valid_account_extfvk(
-        &self,
-        account: AccountId,
-        extfvk: &ExtendedFullViewingKey,
-    ) -> Result<bool, Self::Error>;
-
-    /// Returns the wallet balance for an account as of the specified block
-    /// height.
-    ///
-    /// This may be used to obtain a balance that ignores notes that have been
-    /// received so recently that they are not yet deemed spendable.
-    async fn get_balance_at(
-        &self,
-        account: AccountId,
-        anchor_height: BlockHeight,
-    ) -> Result<Amount, Self::Error>;
-
-    /// Returns the memo for a note.
-    ///
-    /// Implementations of this method must return an error if the note identifier
-    /// does not appear in the backing data store.
-    async fn get_memo(&self, id_note: Self::NoteRef) -> Result<Memo, Self::Error>;
-
-    /// Returns the note commitment tree at the specified block height.
-    async fn get_commitment_tree(
-        &self,
-        block_height: BlockHeight,
-    ) -> Result<Option<CommitmentTree<Node>>, Self::Error>;
-
-    /// Returns the incremental witnesses as of the specified block height.
-    #[allow(clippy::type_complexity)]
-    async fn get_witnesses(
-        &self,
-        block_height: BlockHeight,
-    ) -> Result<Vec<(Self::NoteRef, IncrementalWitness<Node>)>, Self::Error>;
-
-    /// Returns the unspent nullifiers, along with the account identifiers
-    /// with which they are associated.
-    async fn get_nullifiers(&self) -> Result<Vec<(AccountId, Nullifier)>, Self::Error>;
-
-    /// Return all spendable notes.
-    async fn get_spendable_notes(
-        &self,
-        account: AccountId,
-        anchor_height: BlockHeight,
-    ) -> Result<Vec<SpendableNote>, Self::Error>;
-
-    /// Returns a list of spendable notes sufficient to cover the specified
-    /// target value, if possible.
-    async fn select_spendable_notes(
-        &self,
-        account: AccountId,
-        target_value: Amount,
-        anchor_height: BlockHeight,
-    ) -> Result<Vec<SpendableNote>, Self::Error>;
-}
-
-/// This trait encapsulates the write capabilities required to update stored
-/// wallet data.
-#[async_trait::async_trait]
-pub trait WalletWrite: WalletRead {
-    #[allow(clippy::type_complexity)]
-    async fn advance_by_block(
-        &mut self,
-        block: &PrunedBlock,
-        updated_witnesses: &[(Self::NoteRef, IncrementalWitness<Node>)],
-    ) -> Result<Vec<(Self::NoteRef, IncrementalWitness<Node>)>, Self::Error>;
-
-    async fn store_received_tx(
-        &mut self,
-        received_tx: &ReceivedTransaction,
-    ) -> Result<Self::TxRef, Self::Error>;
-
-    async fn store_sent_tx(
-        &mut self,
-        sent_tx: &SentTransaction,
-    ) -> Result<Self::TxRef, Self::Error>;
-
-    /// Rewinds the wallet database to the specified height.
-    ///
-    /// This method assumes that the state of the underlying data store is
-    /// consistent up to a particular block height. Since it is possible that
-    /// a chain reorg might invalidate some stored state, this method must be
-    /// implemented in order to allow users of this API to "reset" the data store
-    /// to correctly represent chainstate as of a specified block height.
-    ///
-    /// After calling this method, the block at the given height will be the
-    /// most recent block and all other operations will treat this block
-    /// as the chain tip for balance determination purposes.
-    ///
-    /// There may be restrictions on how far it is possible to rewind.
-    async fn rewind_to_height(&mut self, block_height: BlockHeight) -> Result<(), Self::Error>;
 }
 
 use crate::error::SqliteClientError;
