+++ zcash_extras/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "zcash_extras"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-trait = "0.1.52"
+zcash_client_backend = { version = "0.5", path = "../zcash_client_backend" }
+zcash_primitives = { version = "0.5", path = "../zcash_primitives" }
